requirements.txt
requests
discord.py
pyfiglet